+++ src/Prbs/Prbs.h
@@ -17,15 +17,15 @@
 class Prbs
 {
 public:
-	Prbs(uint16_t min, uint16_t max, uint16_t seed);
+	Prbs(uint32_t min, uint32_t max, uint16_t seed);
 
 	// return the next value of the sequence
-	uint16_t nextValue();
+	uint32_t nextValue();
 
 private:
 	uint16_t m_reg;
-	uint16_t m_max;
-	uint16_t m_min;
+	uint32_t m_max;
+	uint32_t m_min;
 
 	uint16_t getMem(int index);
